dcp-815 split schema classes
