+++ exporter.py
@@ -10,7 +10,7 @@ from exporter.graph.graph_crawler import GraphCrawler
 from exporter.ingest.service import IngestService
 from exporter.metadata.service import MetadataService
 from exporter.queue.config import AmqpConnConfig, QueueConfig
-from exporter.schema import SchemaService
+from exporter.schema.service import SchemaService
 from exporter.session_context import configure_logger
 from exporter.terra.client import TerraClient
 from exporter.terra.exporter import TerraExporter
