+++ exporter/schema/service.py
@@ -3,25 +3,8 @@ from typing import Optional
 from cachetools.func import ttl_cache
 from hca_ingest.api.ingestapi import IngestApi
 
-
-class SchemaParseException(Exception):
-    pass
-
-
-class SchemaResource:
-
-    def __init__(self, schema_url: str, schema_version: str):
-        self.schema_url = schema_url
-        self.schema_version = schema_version
-
-    @staticmethod
-    def from_dict(data: dict) -> 'SchemaResource':
-        try:
-            schema_url = data["_links"]["json-schema"]["href"]
-            schema_version = data["schemaVersion"]
-            return SchemaResource(schema_url, schema_version)
-        except (KeyError, TypeError) as e:
-            raise SchemaParseException(e)
+from exporter.schema.exceptions import SchemaParseException
+from exporter.schema.resource import SchemaResource
 
 
 class SchemaService:
